+++ src/app/basicauction/basicauction.tsx
@@ -1,51 +1,34 @@
-'use client';
+import { Suspense } from 'react';
 
-import { useSearchParams } from 'next/navigation';
+import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
 
-import useGetBasicAuctionList from '@/apis/queryHooks/basicAuction/useGetBasicAuctionList';
-import AuctionCard from '@/components/AuctionCard';
-import ListLayout from '@/components/ListLayout';
-import { AUCTIONPARAM_KEY } from '@/static/queryParam';
+import { getBasicAuctionList } from '@/apis/queryFunctions/basicAuction';
+import BasicAuctionClientPage from '@/components/BasicAuctionClientPage';
+import AuctionListSkeletonUI from '@/components/SkeletonUI/AuctionListSkeletonUI';
 
-import * as S from './Basicauction.css';
+async function BasicAuction() {
+  const queryClient = new QueryClient();
 
-function BasicAuction() {
-  const searchParams = useSearchParams();
-  const searchKeywordParam = searchParams.get(AUCTIONPARAM_KEY.Q);
-
-  const { data } = useGetBasicAuctionList({
-    title: searchKeywordParam || '',
-    auctionStatus: searchParams.get(AUCTIONPARAM_KEY.AUCTIONSTATUS) || '',
-    sort: searchParams.get(AUCTIONPARAM_KEY.SORT) || '',
-    direction: searchParams.get(AUCTIONPARAM_KEY.DIRECTION) || '',
+  const params = {
+    title: '',
+    auctionStatus: '',
+    sort: '',
+    direction: '',
     page: 0,
     size: 20,
+  };
+
+  await queryClient.prefetchQuery({
+    queryKey: ['basicAuctionList', params],
+    queryFn: () => getBasicAuctionList(params),
   });
 
   return (
-    <div className={S.container}>
-      <div className={S.searchContainer}>
-        <div className={S.count}>
-          <span>전체</span>
-          <span>{data.result_data.content.length}</span>
-        </div>
-      </div>
-      <ListLayout>
-        {data.result_data.content.map(item => (
-          <AuctionCard
-            key={item.auction_id}
-            id={item.auction_id}
-            image={item.image_keys}
-            title={item.title}
-            isLike={false}
-            startPrice={item.start_price}
-            startTime={item.start_date}
-            endTime={item.end_date}
-            nowPrice={item.now_price}
-          />
-        ))}
-      </ListLayout>
-    </div>
+    <HydrationBoundary state={dehydrate(queryClient)}>
+      <Suspense fallback={<AuctionListSkeletonUI count={6} />}>
+        <BasicAuctionClientPage />
+      </Suspense>
+    </HydrationBoundary>
   );
 }
