+++ src/app/basicauction/page.tsx
@@ -1,10 +1,7 @@
-import { Suspense } from 'react';
-
 import BasicAuction from '@/app/basicauction/basicauction';
 import AuctionDropDown from '@/app/basicauction/components/auctiondropdown';
 import LeftSection from '@/app/basicauction/components/leftsection';
 import MaxLayout from '@/components/MaxLayout';
-import AuctionListSkeletonUI from '@/components/SkeletonUI/AuctionListSkeletonUI';
 
 import * as S from './Basicauction.css';
 
@@ -15,9 +12,7 @@ function Home() {
         <LeftSection />
         <section className={S.rightSection}>
           <AuctionDropDown />
-          <Suspense fallback={<AuctionListSkeletonUI count={6} />}>
-            <BasicAuction />
-          </Suspense>
+          <BasicAuction />
         </section>
       </div>
     </MaxLayout>
